Apply all settings in MicroActivity
